policy_hal: avoid unnecessary closure of HOTWORD recording
When the new selected input device is not supported by current
input handle, normally it needs to be closed and then create
a new one to capture with the new device.
While for HOTWORD recording, as the data will either be captured
from handset mic or headset mic, recording can be continued on
the original input handle.

Change-Id: I0f20811effc18433d97b5460b66d4fed21c323b4
Signed-off-by: DennySPB <dennyspb@gmail.com>